Projeto-Integrado-de-Computacao-II/site/requirements.txt
cmake==3.25.2
wheel==0.40.0
dlib==19.24.1
numpy==1.22.4
click==8.1.3
face-recognition==1.3.0

django>=4.2.1
django-stdimage==6.0.1
gunicorn==20.1.0
sqlparse==0.4.4
whitenoise==6.4.0

paho-mqtt==1.6.1
